Refactor MetadataRepository
